int64 -> int; int64 not supported for emitting

+++ metric/counter.go
@@ -26,15 +26,15 @@ func (c *Counter) Dec() {
 	atomic.AddInt64(&c.cur, -1)
 }
 
-func (c *Counter) Max() int64 {
+func (c *Counter) Max() int {
 	cur := atomic.LoadInt64(&c.cur)
 	max := atomic.SwapInt64(&c.max, -1)
 
 	if max == -1 {
 		// no call to .Inc has occurred since last call to .Max;
 		// highest value must be the current value
-		return cur
+		return int(cur)
 	}
 
-	return max
+	return int(max)
 }
